+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/eager/EagerSingletonInstaller.java
@@ -3,6 +3,8 @@ package ru.vyarus.dropwizard.guice.module.installer.feature.eager;
 
 import com.google.common.base.Preconditions;
 import com.google.inject.Binder;
+import com.google.inject.Binding;
+import com.google.inject.spi.DefaultBindingScopingVisitor;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.install.binding.BindingInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.order.Order;
@@ -25,6 +27,8 @@ import java.util.Set;
  */
 @Order(50)
 public class EagerSingletonInstaller implements FeatureInstaller<Object>, BindingInstaller {
+    private static final ScopingVisitor VISITOR = new ScopingVisitor();
+
     private final Reporter reporter = new Reporter(EagerSingletonInstaller.class, "eager singletons =");
     private final Set<String> prerender = new LinkedHashSet<>();
 
@@ -34,9 +38,21 @@ public class EagerSingletonInstaller implements FeatureInstaller<Object>, Bindin
     }
 
     @Override
-    public <T> void install(final Binder binder, final Class<? extends T> type, final boolean lazy) {
+    public void bindExtension(final Binder binder, final Class<?> type, final boolean lazy) {
         Preconditions.checkArgument(!lazy, "Eager bean can't be annotated as lazy: %s", type.getName());
         binder.bind(type).asEagerSingleton();
+    }
+
+    @Override
+    public <T> void checkBinding(final Binder binder, final Class<T> type, final Binding<T> manualBinding) {
+        // we can only validate existing binding here (actually entire extension is pretty useless in case of manual
+        // binding)
+        Preconditions.checkArgument(manualBinding.acceptScopingVisitor(VISITOR),
+                "Eager bean, declared manually is not marked .asEagerSingleton(): %s", type.getName());
+    }
+
+    @Override
+    public void installBinding(final Binder binder, final Class<?> type) {
         // may be called multiple times if bindings report enabled, but log must be counted just once
         prerender.add(String.format("(%s)", type.getName()));
     }
@@ -48,5 +64,20 @@ public class EagerSingletonInstaller implements FeatureInstaller<Object>, Bindin
         }
         prerender.clear();
         reporter.report();
+    }
+
+    /**
+     * Visitor detects eager singleton configuration on binding.
+     */
+    private static class ScopingVisitor extends DefaultBindingScopingVisitor<Boolean> {
+        @Override
+        protected Boolean visitOther() {
+            return false;
+        }
+
+        @Override
+        public Boolean visitEagerSingleton() {
+            return true;
+        }
     }
 }