+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/jersey/ResourceInstaller.java
@@ -1,7 +1,9 @@
 package ru.vyarus.dropwizard.guice.module.installer.feature.jersey;
 
 
+import com.google.common.base.Preconditions;
 import com.google.inject.Binder;
+import com.google.inject.Binding;
 import com.google.inject.Injector;
 import io.dropwizard.setup.Environment;
 import org.glassfish.jersey.internal.inject.AbstractBinder;
@@ -42,13 +44,27 @@ public class ResourceInstaller extends AbstractJerseyInstaller<Object> implement
     }
 
     @Override
-    public <T> void install(final Binder binder, final Class<? extends T> type, final boolean lazyMarker) {
+    public void bindExtension(final Binder binder, final Class<?> type, final boolean lazyMarker) {
         final boolean jerseyManaged = isJerseyExtension(type);
         final boolean lazy = isLazy(type, lazyMarker);
         // register in guice only if not managed by hk and just in time (lazy) binding not requested
         if (!jerseyManaged && !lazy) {
             bindInGuice(binder, type);
         }
+    }
+
+    @Override
+    public <T> void checkBinding(final Binder binder, final Class<T> type, final Binding<T> manualBinding) {
+        // no need to bind in case of manual binding
+        final boolean hkManaged = isJerseyExtension(type);
+        Preconditions.checkState(!hkManaged,
+                "Resource annotated as jersey managed is declared manually in guice: %s",
+                type.getName());
+    }
+
+    @Override
+    public void installBinding(final Binder binder, final Class<?> type) {
+        // no common binding logic (reporting is handled by dropwizard)
     }
 
     @Override