+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/order/OrderedApplication.groovy
@@ -1,5 +1,7 @@
 package ru.vyarus.dropwizard.guice.support.order
 
+import com.google.inject.Binder
+import com.google.inject.Module
 import io.dropwizard.Application
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
@@ -7,7 +9,6 @@ import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.support.TestConfiguration
 import ru.vyarus.dropwizard.guice.support.feature.DummyResource
 import ru.vyarus.dropwizard.guice.support.feature.DummyService
-import ru.vyarus.dropwizard.guice.support.util.BindModule
 
 /**
  * Application to check extension ordering.
@@ -23,7 +24,12 @@ class OrderedApplication extends Application<TestConfiguration> {
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.order")
                 // at least one resource required
                 .extensions(DummyResource)
-                .modules(new BindModule(DummyService))
+                .modules(new Module() {
+                    @Override
+                    void configure(Binder binder) {
+                        binder.bind(DummyService).asEagerSingleton()
+                    }
+                })
                 .build()
         );
     }