+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/web/ServletsApplication.groovy
@@ -1,12 +1,13 @@
 package ru.vyarus.dropwizard.guice.support.web
 
+import com.google.inject.Binder
+import com.google.inject.Module
 import io.dropwizard.Application
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.support.TestConfiguration
 import ru.vyarus.dropwizard.guice.support.feature.DummyService
-import ru.vyarus.dropwizard.guice.support.util.BindModule
 
 /**
  * @author Vyacheslav Rusakov 
@@ -18,7 +19,12 @@ class ServletsApplication extends Application<TestConfiguration> {
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
         bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.web.feature")
-                .modules(new WebModule(), new BindModule(DummyService))
+                .modules(new WebModule(), new Module() {
+                    @Override
+                    void configure(Binder binder) {
+                        binder.bind(DummyService).asEagerSingleton()
+                    }
+                })
                 .build()
         );
     }